table: fix bus error by substituting float with double.

+++ include/rime/dict/table.h
@@ -29,7 +29,7 @@ typedef List<SyllableId> Code;
 
 struct Entry {
   String text;
-  float weight;
+  double weight;
 };
 
 struct HeadIndexNode {
